fix(UI-fixes):- all-in amount changed to float, deselection of auto-options, i-am-back- button display only on sitout, auto-rotation added to settings and some more UI fixes related to portrait mode

+++ app/src/main/java/com/gtgt/pokerjacks/ui/game/models/GameModel.kt
@@ -113,7 +113,7 @@ data class PlayerTurn(
     val player_grace_timer: Long,
     val player_turn: String,
     val current_bettor_position: Int,
-    val allin_amount: Int
+    val allin_amount: Double
 //    val total_pot_value: Double
 )
